Extract add and delete into packages. Create stubby FindInternshipCommand.
