+++ src/main/java/seedu/address/logic/commands/FindInternshipCommand.java
@@ -1,0 +1,9 @@
+package seedu.address.logic.commands;
+
+import seedu.address.model.item.field.NameContainsKeywordsPredicate;
+
+public class FindInternshipCommand extends FindCommand {
+    public FindInternshipCommand(NameContainsKeywordsPredicate predicate) {
+        super(predicate);
+    }
+}
