+++ src/main/java/seedu/address/logic/commands/add/AddCommand.java
@@ -1,4 +1,6 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.add;
+
+import seedu.address.logic.commands.Command;
 
 import static seedu.address.logic.parser.CliSyntax.PREFIX_ADDRESS;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_EMAIL;
