+++ src/main/java/seedu/address/logic/commands/delete/DeleteSkillCommand.java
@@ -1,19 +1,20 @@
-package seedu.address.logic.commands;
-
-import static java.util.Objects.requireNonNull;
+package seedu.address.logic.commands.delete;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.item.Internship;
+import seedu.address.model.item.Skill;
+
+import static java.util.Objects.requireNonNull;
 
 /**
- * Deletes an Internship item.
+ * Deletes a Skill item.
  */
-public class DeleteInternshipCommand extends DeleteCommand {
+public class DeleteSkillCommand extends DeleteCommand {
 
-    public DeleteInternshipCommand(Index targetIndex) {
+    public DeleteSkillCommand(Index targetIndex) {
         super(targetIndex);
     }
 
@@ -21,14 +22,14 @@ public class DeleteInternshipCommand extends DeleteCommand {
     public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
 
-        if (targetIndex.getZeroBased() >= model.getResumeSize()) {
+        if (targetIndex.getZeroBased() >= model.getSkillSize()) {
             throw new CommandException(Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
         }
 
-        Internship toDelete = model.getInternship(targetIndex);
+        Skill toDelete = model.getSkill(targetIndex);
 
-        model.deleteInternship(toDelete);
-        model.setInternshipToDisplay();
+        model.deleteSkill(toDelete);
+        model.setSkillToDisplay();
 
         return new CommandResult(toDelete.toString(),
                 String.format(MESSAGE_DELETE_ITEM_SUCCESS, toDelete));
@@ -37,7 +38,7 @@ public class DeleteInternshipCommand extends DeleteCommand {
     @Override
     public boolean equals(Object other) {
         return other == this // short circuit if same object
-                || (other instanceof DeleteInternshipCommand // instanceof handles nulls
-                && targetIndex.equals(((DeleteInternshipCommand) other).targetIndex)); // state check
+                || (other instanceof DeleteSkillCommand // instanceof handles nulls
+                && targetIndex.equals(((DeleteSkillCommand) other).targetIndex)); // state check
     }
 }
