+++ src/main/java/seedu/address/logic/parser/AddCommandParser.java
@@ -12,9 +12,9 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_TO;
 import java.util.Set;
 import java.util.stream.Stream;
 
-import seedu.address.logic.commands.AddCommand;
-import seedu.address.logic.commands.AddInternshipCommand;
-import seedu.address.logic.commands.AddResumeCommand;
+import seedu.address.logic.commands.add.AddCommand;
+import seedu.address.logic.commands.add.AddInternshipCommand;
+import seedu.address.logic.commands.add.AddResumeCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.item.Internship;
 import seedu.address.model.item.Item;
