+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -17,7 +17,7 @@ package seedu.address.logic;
 //import org.junit.jupiter.api.Test;
 //import org.junit.jupiter.api.io.TempDir;
 //
-//import seedu.address.logic.commands.AddCommand;
+//import seedu.address.logic.commands.add.AddCommand;
 //import seedu.address.logic.commands.CommandResult;
 //import seedu.address.logic.commands.ListCommand;
 //import seedu.address.logic.commands.exceptions.CommandException;
