+++ src/test/java/seedu/address/logic/commands/DeleteCommandTest.java
@@ -14,6 +14,8 @@ import org.junit.jupiter.api.Test;
 /*import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;*/
 
+import seedu.address.logic.commands.delete.DeleteCommand;
+import seedu.address.logic.commands.delete.DeleteInternshipCommand;
 import seedu.address.model.Model;
 
 /*import seedu.address.model.ModelManager;
